Extract AccountButtons and CategoryButtons components
